Added enum for ConnectionStatus for a CDevice. Fixed a bug in make_uuid() that caused windows guid to not be converted correctly.
Added struct header for whereami's write characterisitic - demand rssi.

CAuthenticator.hpp now owns CDevice connection to the authenticated server. Added deauthentication function to CAuthenticator.

CCharacteristic now catches potential write exceptions thrown by winrt.

modified CCharacterisitic to return CommunicationStatus instead of winrt specific enum.

Added callback to CDevice that triggers on connection changed event.

Added destructor to CService since winrt doesn't properly close the handles when there are no strong references left (?).

+++ client/src/gui/CGui.cpp
@@ -240,7 +240,7 @@ CGui::CGui()
     load_layout();
     
     ImGuiIO& io = ImGui::GetIO();
-    io.Fonts->AddFontFromMemoryTTF(FontLouisGeorge.data(), FontLouisGeorge.size(), 18);
+    io.Fonts->AddFontFromMemoryTTF(FontLouisGeorge.data(), static_cast<int32_t>(FontLouisGeorge.size()), 18);
 }
 CGui::~CGui()
 {
